fix(storage): fix PostgreSQL adapter compilation errors
- Fix close() error handling in SSL connection test (adapter.v:59)
- Fix bit shift warning for signed int in hex parsing (cluster_metadata.v:656)
- Add proper error handling with 'or {}' block
- Cast to u8 before bit shift to avoid sign issues

All tests now pass successfully.

+++ src/infra/storage/plugins/postgres/cluster_metadata.v
@@ -653,7 +653,7 @@ fn parse_hex_string(s string) []u8 {
 		high := hex_char_to_nibble(s[i])
 		low := hex_char_to_nibble(s[i + 1])
 		if high >= 0 && low >= 0 {
-			result << u8((high << 4) | low)
+			result << u8((u8(high) << 4) | u8(low))
 		}
 		i += 2
 	}
